Extract common experiments functions to experiments/functions
