+++ hgrlab/experiments/delello_lnlm2024.py
@@ -1,16 +1,14 @@
 import numpy as np
-import multiprocessing as mp
-import concurrent.futures
-import datetime
 
+from ..utils import AssetManager
 from ..models.hgrdtw import k_fold_cost, FeatureSet
 from ..models.hgrdtw import build_classifier, fit, predict
-from ..utils import AssetManager
 
-from ..experiments import run_experiments, print_message, print_progress, print_result, print_line_break, download_assets
 from ..experiments.data import emgepn10
+from ..experiments import run_experiments, download_assets
+from .functions import eval_hgr_systems, tune_seg_thresholds
 
-def find_optimum_segmentation_threshold(config):
+def tune_segmentation_threshold(config):
     threshold_min = config['threshold_min']
     threshold_max = config['threshold_max']
     threshold_direction = config['threshold_direction']
@@ -51,134 +49,6 @@ def find_optimum_segmentation_threshold(config):
 
     return thresholds[np.argmax(thresholds_errors)]
 
-def find_optimum_segmentation_thresholds_by_classifier_and_user(
-    experiment_id,
-    total_experiments,
-    dataset_name,
-    assets_dir,
-    user_ids,
-    options,
-    threshold_min=10,
-    threshold_max=20,
-):
-    start_ts = datetime.datetime.now()
-
-    classifier_names = options['classifier_names']
-    folds = options['cv_folds']
-
-    feature_window_length = options['feature_window_length']
-    feature_overlap_length = options['feature_overlap_length']
-
-    if 'cv_options' in options.keys():
-        cv_options = options['cv_options']
-    else:
-        cv_options = None
-
-    task = 'Optimizing thresholds'
-
-    print_message(
-        'Experiment %d of %d: Optimize HGR segmentation thresholds using %d-fold cross-validation' % (
-            experiment_id,
-            total_experiments,
-            folds,
-    ))
-    print_message('Classifiers: %s' % classifier_names)
-    print_message('Number of subjects: %d' % np.size(user_ids))
-
-    optimum_thresholds = np.zeros((np.size(classifier_names), np.size(user_ids)), dtype=int)
-
-    num_workers = mp.cpu_count()
-
-    def get_progress(classifier_id, config_id=0):
-        current = np.size(user_ids) * classifier_id + config_id
-        total = np.size(user_ids) * np.size(classifier_names)
-        return current / total
-
-    for classifier_id, classifier_name in enumerate(classifier_names):
-        print_line_break()
-        print_progress(
-            task,
-            get_progress(classifier_id),
-            'running %s-fold cross-validation on classifier %s...' % (
-                folds,
-                classifier_name,
-            )
-        )
-
-        user_configs = []
-
-        for user_id in user_ids:
-            config = {
-                'classifier_name': classifier_name,
-                'threshold_min': threshold_min,
-                'threshold_max': threshold_max,
-                'threshold_direction': 'desc',
-                'cv_folds': folds,
-                'cv_options': cv_options,
-                'feature_set_config': {
-                    'user_id': user_id,
-                    'ds_name': dataset_name,
-                    'ds_type': 'training',
-                    'ds_dir': assets_dir,
-                    'fs_dir': assets_dir,
-                    'stft_window_length': 25,
-                    'stft_window_overlap': 10,
-                    'stft_nfft': 50,
-                    'activity_extra_samples': 25,
-                    'activity_min_length': 100,
-                    'feature_window_length': feature_window_length,
-                    'feature_overlap_length': feature_overlap_length,
-                },
-            }
-            
-            user_configs.append(config)
-        
-        with concurrent.futures.ProcessPoolExecutor(max_workers=num_workers) as executor:
-            for i, config, result in zip(
-                np.arange(0, np.size(user_configs)),
-                user_configs,
-                executor.map(
-                    find_optimum_segmentation_threshold,
-                    user_configs,
-                )
-            ):
-                print_progress(
-                    task,
-                    get_progress(classifier_id, i+1),
-                    'optimized classifier %s for subject %2d of %2d' % (
-                        classifier_name,
-                        i+1,
-                         np.size(user_configs),
-                    )
-                )
-                
-                optimum_thresholds[classifier_id,i] = result
-
-    end_ts = datetime.datetime.now()
-
-    output_message = '%s\n%s\n' % (
-        'Optimum individual segmentation thresholds using 4-fold cross-validation',
-        'Lines: classifiers | Columns: subjects'
-    )
-    for classifier_id, classifier in enumerate(classifier_names):
-        output_message = output_message + '\n%03s: %s' % (
-            classifier,
-            optimum_thresholds[classifier_id],
-        )
-    
-    print_line_break()
-    print_message('Finished segmentation threshold optimization')
-    print_message('Time elapsed in experiment %d of %d: %s' % (
-        experiment_id,
-        total_experiments,
-        str(end_ts - start_ts),
-    ))
-
-    return {
-        'data': optimum_thresholds,
-        'message': output_message,
-    }
-
 def reduce_window_predictions(
     predictions,
     window_start_activity_indexes,
@@ -215,7 +85,7 @@ def reduce_window_predictions(
 
     return reduced_prediction
 
-def assess_hand_gesture_classification(config):
+def eval_hgr_system(config):
     experiments = config['experiments']
     classifier_name = config['classifier_name']
 
@@ -257,173 +127,44 @@ def assess_hand_gesture_classification(config):
 
     return errors, trials
 
-def assess_hgr_systems_by_classifier_and_user(
+def tune_segmentation_thresholds_by_classifier_and_user(
     experiment_id,
     total_experiments,
     dataset_name,
     assets_dir,
     user_ids,
     options,
-    experiment_runs=100,
 ):
-    start_ts = datetime.datetime.now()
-
-    classifier_names = options['classifier_names']
-    thresholds = options['thresholds']
-    feature_window_length = options['feature_window_length']
-    feature_overlap_length = options['feature_overlap_length']
-
-    task = 'Assessing HGR systems'
-
-    max_workers = mp.cpu_count()
-    number_of_users = np.size(user_ids)
-    number_of_classifiers = np.size(classifier_names)
-
-    errors = np.zeros(
-        (number_of_classifiers,number_of_users,experiment_runs,),
-        dtype=np.uint32
-    )
-
-    trials = np.zeros(
-        (number_of_classifiers,number_of_users,experiment_runs,),
-        dtype=np.uint32
-    )
-
-    print_message(
-        'Experiment %d of %d: Compare the accuracy of HGR systems using %d classifiers ' % (
-            experiment_id,
-            total_experiments,
-            number_of_classifiers,
-    ))
-    print_message('Classifiers: %s' % classifier_names)
-    print_message('Number of subjects: %d' % number_of_users)
-
-    def get_progress(classifier_id, config_id=0):
-        current = number_of_users * classifier_id + config_id
-        total = number_of_users * number_of_classifiers
-        return current / total
-    
-    for classifier_id, classifier_name in enumerate(classifier_names):
-        print_line_break()
-        print_progress(
-            task,
-            get_progress(classifier_id),
-            'evaluating classifier %s...' % classifier_name
-        )
-
-        user_configs = []
-
-        for i, user_id in enumerate(user_ids):
-            optimum_threshold = thresholds[classifier_name][i]
-            config = {
-                'classifier_name': classifier_name,
-                'experiments': experiment_runs,
-                'feature_set_config': {
-                    'user_id': user_id,
-                    'ds_name': dataset_name,
-                    'ds_dir': assets_dir,
-                    'fs_dir': assets_dir,
-                    'stft_window_length': 25,
-                    'stft_window_overlap': 10,
-                    'stft_nfft': 50,
-                    'activity_threshold': optimum_threshold,
-                    'activity_extra_samples': 25,
-                    'activity_min_length': 100,
-                    'feature_window_length': feature_window_length,
-                    'feature_overlap_length': feature_overlap_length,
-                },
-            }
-            
-            user_configs.append(config)
-
-        with concurrent.futures.ProcessPoolExecutor(max_workers=max_workers) as executor:
-            for i, config, result in zip(
-                np.arange(0, np.size(user_configs)),
-                user_configs,
-                executor.map(
-                    assess_hand_gesture_classification,
-                    user_configs,
-                )
-            ):
-                print_progress(
-                    task,
-                    get_progress(classifier_id, i+1),
-                    'evaluated classifier %s for subject %2d of %2d' % (
-                        classifier_name,
-                        i+1,
-                        np.size(user_configs),
-                    )
-                )
-                
-                errors[classifier_id,i,:] = result[0]
-                trials[classifier_id,i,:] = result[1]
-
-    accuracy = (trials - errors) / trials
-    accuracy_mean = accuracy.mean(axis=(1,2))
-    accuracy_per_experiment = accuracy.mean(axis=1)
-    accuracy_per_user = accuracy.mean(axis=2)
-
-    end_ts = datetime.datetime.now()
-
-    acc_by_classifier_output = ''
-    for classifier_id, classifier in enumerate(classifier_names):
-        acc_by_classifier_output = '%s\n%03s: %.1f \u00B1 %.1f' % (
-            acc_by_classifier_output,
-            classifier,
-            accuracy_mean[classifier_id] * 100,
-            accuracy_per_experiment[classifier_id].std(ddof=1) * 100,
-        )
-
-    acc_by_classifier_output = '%s\n%s\n%s' % (
-        'Mean accuracy and standard deviation of the HGR systems using different classifiers',
-        'Lines: classifiers | Data: mean accuracy and standard deviation (%)',
-        acc_by_classifier_output,
-    )
-    
-    acc_by_subject_output = '    ' + '    '.join(
-        ['%12s' % name for name in classifier_names]
-    )
-
-    for i, user_id in enumerate(user_ids):
-        acc_by_subject_output = acc_by_subject_output + '\n#%02s:' % user_id
-        for classifier_id, classifier in enumerate(classifier_names):
-            acc_by_subject_output = '{OUTPUT}    {MEAN:5.1f} \u00B1 {STD:4.1f}'.format(
-                OUTPUT=acc_by_subject_output,
-                MEAN=accuracy_per_user[classifier_id,i] * 100,
-                STD=accuracy[classifier_id,i].std(ddof=1) * 100,
-            )
-
-    acc_by_subject_output = acc_by_subject_output + '\n\nAVG '
-    for classifier_id, classifier in enumerate(classifier_names):
-        acc_by_subject_output = '{OUTPUT}    {MEAN:5.1f} \u00B1 {STD:4.1f}'.format(
-            OUTPUT=acc_by_subject_output,
-            MEAN=accuracy_mean[classifier_id] * 100,
-            STD=accuracy_per_user.std(axis=1,ddof=1)[classifier_id] * 100,
-        )
-
-    acc_by_subject_output = '%s\n%s\n%s' % (
-        'Mean accuracy and standard deviation by subject for different classifiers',
-        'Lines: subjects | Columns: classifiers | Data: mean accuracy and standard deviation (%)',
-        acc_by_subject_output,
-    )
-
-    output_message = '%s\n\n%s' % (
-        acc_by_classifier_output,
-        acc_by_subject_output,
-    )
-    
-    print_line_break()
-    print_message('Finished evaluation of HGR systems')
-    print_message('Time elapsed in experiment %d of %d: %s' % (
+    return tune_seg_thresholds.run(
         experiment_id,
         total_experiments,
-        str(end_ts - start_ts),
-    ))
+        dataset_name,
+        assets_dir,
+        user_ids,
+        options,
+        threshold_min=10,
+        threshold_max=20,
+        tune_segmentation_threshold=tune_segmentation_threshold,
+    )
 
-    return {
-        'data': accuracy,
-        'message': output_message,
-    }
+def eval_hgr_systems_by_classifier_and_user(
+    experiment_id,
+    total_experiments,
+    dataset_name,
+    assets_dir,
+    user_ids,
+    options,
+):
+    return eval_hgr_systems.run(
+        experiment_id,
+        total_experiments,
+        dataset_name,
+        assets_dir,
+        user_ids,
+        options,
+        experiment_runs=100,
+        eval_hgr_system=eval_hgr_system,
+    )
 
 def main():
     authors = 'Guilherme C. De Lello, Gabriel S. Chaves, Juliano F. Caldeira, and Markus V.S. Lima'
@@ -444,8 +185,8 @@ def main():
         user_ids=np.arange(1, 11),
         setup=setup,
         experiments=[
-            find_optimum_segmentation_thresholds_by_classifier_and_user,
-            assess_hgr_systems_by_classifier_and_user,
+            tune_segmentation_thresholds_by_classifier_and_user,
+            eval_hgr_systems_by_classifier_and_user,
         ],
         options={
             'cv_folds': 4,
